requirements.txt
pandas
requests
scipy
pingouin
censys
google-cloud-bigquery
db-dtypes